+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Change Log
 
-## 0.1.2 (unreleased)
+## 0.2 (unreleased)
 
 * Added all of the resources and resource properties for Auto Scaling Groups.
+* New AutoScalingGroup example
+* Added UpdatePolicy, CreationPolicy, and DependsOn
 
 ## 0.1.1